+++ CHANGELOG.md
@@ -26,6 +26,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 ### Added
 
+-   [#142](https://github.com/ARGOeu/argo-accounting/pull/142) - Enhance Metric Response.
 -   [#141](https://github.com/ARGOeu/argo-accounting/pull/141) - Accounting Service health checks.
 -   [#138](https://github.com/ARGOeu/argo-accounting/pull/138) - Docusaurus Guideline for managing Metric Definitions via UI.